Rename adapter_connector → transfer_context
